+++ tarantool/src/lib.rs
@@ -75,6 +75,7 @@ pub mod uuid;
 #[cfg(all(target_arch = "aarch64", target_os = "macos"))]
 #[doc(hidden)]
 mod va_list;
+pub mod vclock;
 
 /// `#[tarantool::proc]` is a macro attribute for creating stored procedure
 /// functions.